Move files to overall package structure
