+++ reproductive_rights_data_project/__main__.py
@@ -4,9 +4,9 @@ packages based on the flags that are passed to it.
 """
 
 import argparse
-import visualization.service
-import api.service
-import data_handling.service
+import reproductive_rights_data_project.visualization.service
+import reproductive_rights_data_project.api.service
+import reproductive_rights_data_project.data_handling.service
 
 if __name__ == "__main__":
     """
@@ -39,10 +39,10 @@ if __name__ == "__main__":
     args = parser.parse_args()
 
     if args.api:
-        api.service.main()
+        reproductive_rights_data_project.api.service.main()
 
     if args.parse_data:
-        data_handling.service.main()
+        reproductive_rights_data_project.data_handling.service.main()
 
     if args.visualize:
-        visualization.service.main()
+        reproductive_rights_data_project.visualization.service.main()
