+++ reproductive_rights_data_project/test/test_util.py
@@ -1,8 +1,8 @@
 """
 This file contains all testing functions for the util package.
 """
-from util.constants import STANDARD_ENCODING
-from util.functions import to_json
+from reproductive_rights_data_project.util.constants import STANDARD_ENCODING
+from reproductive_rights_data_project.util.functions import to_json
 import json
 import os
