+++ reproductive_rights_data_project/visualization/maps/usa_country.py
@@ -6,8 +6,9 @@ import plotly.express as px
 import pandas as pd
 import json
 
-from visualization.abstract_visualization import Visualization
-from util.constants import STANDARD_ENCODING
+from reproductive_rights_data_project.visualization.abstract_visualization import \
+    Visualization
+from reproductive_rights_data_project.util.constants import STANDARD_ENCODING
 
 
 class USAMap(Visualization):
@@ -19,10 +20,10 @@ class USAMap(Visualization):
     """
 
     def __init__(
-        self,
-        gestational_info_file_name,
-        locations_file_name,
-        state_abbrevs_file_name,
+            self,
+            gestational_info_file_name,
+            locations_file_name,
+            state_abbrevs_file_name,
     ):
         """
         Author(s): Michael Plunkett
@@ -43,13 +44,12 @@ class USAMap(Visualization):
         """
 
         with open(
-            self._gestational_info_file_name, encoding=STANDARD_ENCODING
+                self._gestational_info_file_name, encoding=STANDARD_ENCODING
         ) as gestational, open(
             self._locations_file_name, encoding=STANDARD_ENCODING
         ) as locations, open(
             self._state_abbrevs_file_name, encoding=STANDARD_ENCODING
         ) as abbreviations:
-
             self._gestational_info = json.load(gestational)
             self._locations = json.load(locations)
             self._state_abbrevs = pd.read_csv(abbreviations)
